Rename ShopMatcher to POIMatcher
No change in behavior.
